Move Hey func into commands
