+++ slack.go
@@ -5,15 +5,9 @@ import (
 	"log"
 	"strings"
 
+	"github.com/myrefer/go-slack-interactive/commands"
 	"github.com/myrefer/go-slack-interactive/slack"
 	api "github.com/nlopes/slack"
-)
-
-const (
-	// action is used for slack attament action.
-	actionSelect = "select"
-	actionStart  = "start"
-	actionCancel = "cancel"
 )
 
 type SlackListener struct {
@@ -56,72 +50,8 @@ func (s *SlackListener) handleMessageEvent(ev *api.MessageEvent) error {
 
 	// Parse message
 	mux := slack.NewServeMux()
-	mux.Handle("hey", slack.HandlerFunc(Hey))
+	mux.Handle("hey", slack.HandlerFunc(commands.Hey))
 	mux.ServeMessage(ev, s.client)
 
 	return nil
-}
-
-func Hey(ev *api.MessageEvent, client *api.Client) {
-	log.Printf("Start Hey")
-	// value is passed to message handler when request is approved.
-	attachment := api.Attachment{
-		Text:       "Which beer do you want? :beer:",
-		Color:      "#f9a41b",
-		CallbackID: "beer",
-		Actions: []api.AttachmentAction{
-			{
-				Name: actionSelect,
-				Type: "select",
-				Options: []api.AttachmentActionOption{
-					{
-						Text:  "Asahi Super Dry",
-						Value: "Asahi Super Dry",
-					},
-					{
-						Text:  "Kirin Lager Beer",
-						Value: "Kirin Lager Beer",
-					},
-					{
-						Text:  "Sapporo Black Label",
-						Value: "Sapporo Black Label",
-					},
-					{
-						Text:  "Suntory Malts",
-						Value: "Suntory Malts",
-					},
-					{
-						Text:  "Yona Yona Ale",
-						Value: "Yona Yona Ale",
-					},
-					{
-						Text:  "Plemiun Molt",
-						Value: "Plemiun Molt",
-					},
-					{
-						Text:  "Yebisu",
-						Value: "Yebisu",
-					},
-				},
-			},
-
-			{
-				Name:  actionCancel,
-				Text:  "Cancel",
-				Type:  "button",
-				Style: "danger",
-			},
-		},
-	}
-
-	params := api.PostMessageParameters{
-		Attachments: []api.Attachment{
-			attachment,
-		},
-	}
-
-	if _, _, err := client.PostMessage(ev.Channel, "", params); err != nil {
-		log.Printf("failed to post message: %s", err)
-	}
-	log.Printf("Succeeded to post")
 }
